m/n/c/curator: add CuratorLocal.GetCurrentLeader
This adds a service (CuratorLocal) which runs on both leader and
follower curators. It has one RPC, GetCurrentLeader, which returns
information about the leader election status from the point of view of
the callee.

We add a test to make sure the current leader returns correct data, but
we don't yet have a test for a follower (that would require a
significant test harness). In an upcoming CL we'll be exercising this in
an end-to-end test, however.

Change-Id: I4dea780953bdc196bbc5a744f49ee688327c3269
Reviewed-on: https://review.monogon.dev/c/monogon/+/784
Tested-by: Jenkins CI
Reviewed-by: Mateusz Zalega <mateusz@monogon.tech>