m/proto: Add RunningCurator to status, report in status pusher
This data allows more dynamic reporting of a node's Curator status, and
notably allows reporting which port it's running on.

We weren't planning on supporting running on non-standard ports, and we
probably still don't, but it's actually super useful to have this
ability in (future) tests.

We use the opportunity to refactor the roleserver's statuspush worker,
and to add a test for it.

Change-Id: I53322e6c8d268186ede085d4a05b646acb422a6b
Reviewed-on: https://review.monogon.dev/c/monogon/+/793
Reviewed-by: Mateusz Zalega <mateusz@monogon.tech>
Tested-by: Jenkins CI