m/n/c/rpc: make resolver leader-aware
This is a fairly large change which makes the resolver only contact the
current leader of the control plane, not all nodes in a round-robin
fashion.

This resolver isn't yet used by the cluster, the tests, or metroctl -
but that will come in upcoming CLs.

We also move the resolver to a subpackage of rpc, in preparation for
moving it into a package path designed to be depended upon by users.

Change-Id: I230853bbf552fbde947de05f95de37cea93a168c
Reviewed-on: https://review.monogon.dev/c/monogon/+/795
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI