github: implement release workflow
This adds a GitHub Action which builds all release artifacts (currently
metroctl, the bundle and the handbook) and pushes them to a GCS bucket
as well as to GitHub releases. It uses federated identity to
authenticate to Google Cloud via an OIDC token issued by GitHub.
The action only runs for tags starting with a 'v' as it's not intended
as a general-purpose CI solution.

All artifacts are uploaded to GCS even though initially only the
handbook is used there in anticipation of automated downloads.

Change-Id: I125166b2178d5a0aa53b725259ac96386f0275d0
Reviewed-on: https://review.monogon.dev/c/monogon/+/810
Tested-by: Jenkins CI
Reviewed-by: Mateusz Zalega <mateusz@monogon.tech>