m/n/build: fix fsspec rules duplicating specs
Currently if a rule which outputs an fsspec outputs both a DefaultInfo
as well as an FSSpecInfo provider and both are propulated, the fsspec
core implementation will use both instead of only using the FSSpecInfo.
This leads to spurious errors when the fsspec consumers detect duplicate
files.
This change makes the fsspec rule ignore the DefaultInfo provider if a
FSSpecInfo provider exists on an input label.

Change-Id: Iec8975200380889732c840ff16d17f791fa1ba83
Reviewed-on: https://review.monogon.dev/c/monogon/+/902
Tested-by: Jenkins CI
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>