third_party/linux: improve kernel configuration
These configuration changes are required to boot the Metropolis kernel
via kexec on a PowerEdge R750.

kexec needs to be enabled even if we're not using it as just using kexec
to launch a kernel on an EFI machine requires a fixup performed only if
the target kernel is built with kexec enabled. Otherwise it crashes
immediately after mounting efivars by dereferencing a null pointer.

bpfilter should be disabled because it needs a userspace helper (the
.ko actually runs in userspacee) and we ship none of that, causing an
error-level log message on every boot.
Until we actually ship the required infrastructure disable it.

irq_remap is required for x2apic, without it
the kernel can't even boot on that platform.

intel_iommu is just a drive-by enable because the AMD IOMMU is already
enabled and we want the protection.

Change-Id: Iaf0012e8c0427114c56fc5d90a9748ebeb800a54
Reviewed-on: https://review.monogon.dev/c/monogon/+/904
Tested-by: Jenkins CI
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>