pkg/smbios: fix error calculating memory size
There were two errors in the memory size calculation.
One was that if the high bit is set the unit is KiB, otherwise MiB, not
the other way around.
The other one was that I extracted the bit, but failed to shift it to
the right position. So I took the time to stop this bit twiddling and
added some constants and a good old-fashioned if.

Change-Id: I0bce8f7607981e62120365374458425f3c663b51
Reviewed-on: https://review.monogon.dev/c/monogon/+/1000
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>
Tested-by: Jenkins CI