metropolis/test/launch: convert :launch to a test
"bazel run" simply executes the binary outside the sandbox, so swtpm
and other dependencies from the sandbox sysroot won't be available.

If swtpm is installed on the host, running the _bin target still works,
but it's better to point contributors to something that works
out of the box.

This is a temporary workaround. Tests have timeouts and take the global
Bazel server lock. The correct solution is a static swtpm build
which can run outside the sandbox.

Change-Id: Icf7bf5cc44825df676d37a75ea9c1e135de14fef
Reviewed-on: https://review.monogon.dev/c/monogon/+/1078
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>
Tested-by: Jenkins CI