*: use a vishvananda/netlink fork by Monogon
This switches Metropolis over to a Monogon-maintained fork of
vishvananda/netlink.

The package was modified to support the 'sample' action in packet
filters. This is required for the upcoming change enabling network
analytics through local packet capture.

Metropolis' dhcp4c had to be modified to match the vishvananda/netlink
version referenced by this change.

Change-Id: I2dd0799a009618f8543904252b85ff63ddd560c7
Reviewed-on: https://review.monogon.dev/c/monogon/+/1109
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>