cloud: rename ShepherdInstall to ShepherdAccess and clean up other 'installation' references
ShepherdInstall should've been called ShepherdAgentStart from the
beginning, but got named incorrectly because of a momentary lapse of
reason.

Instead of doing a rename, we widen that Process to anything that
involved the Shepherds directly accessing the underlying provider
machines, be it starting the agent, trying to fix the machine, or
rebooting a machine. This wide meaning makes sure no two shepherds work
on the same machine ever, regardless of the actual workflow each one is
performing.

Change-Id: Ic247919d1bcf1c5ec9fcf7125f17b90413068ed5
Reviewed-on: https://review.monogon.dev/c/monogon/+/1138
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>