net/proto: add network configuration spec
This adds a network configuration specification which will be used for
the Monogon Network component in case autoconfiguration cannot be used
because of environmental constraints.

It will also be used by netdump, which is capable of deriving such a
spec from a running system.

It's currently a bit bare-bones, but is designed to be easily extensible to
more complex needs.

Change-Id: Ib444ac26251884c640a0dbec089d0d1787df556b
Reviewed-on: https://review.monogon.dev/c/monogon/+/1139
Reviewed-by: Leopold Schabel <leo@monogon.tech>
Tested-by: Leopold Schabel <leo@monogon.tech>