c/takeover: init
Add takeover, a tool which is used to take over machines running any
pretty much any Linux distribution as long as it has kexec support and
one can run the takeover binary on it.

After takeover it launches the cloud agent which can then perform
further work on the machine.

Change-Id: If26015f626f439c44be473221c98b5e9a8fa9adc
Reviewed-on: https://review.monogon.dev/c/monogon/+/1143
Reviewed-by: Serge Bazanski <serge@monogon.tech>
Tested-by: Leopold Schabel <leo@monogon.tech>