t/go: add netlink patch for permanent HW address
This adds a patch to the netlink dependency which adds a field
containing the permanent hardware address to the Link struct which
can be used to reliably identify interface hardware addresses even
if they have been changed for operational reasons.

The patch has already been sent upstream at
https://github.com/vishvananda/netlink/pull/850 but has not been
reviewed or accepted.

Change-Id: I9562d32643f5b3a5dcf9f1930b951b75ccc30da1
Reviewed-on: https://review.monogon.dev/c/monogon/+/1228
Tested-by: Leopold Schabel <leo@monogon.tech>
Reviewed-by: Serge Bazanski <serge@monogon.tech>