go/types/mapsets: add OrderedMap
Oh look, more Go generics.

This is a pretty standard Ordered/SortedMap type structure, with a
little twist: it implements having a subordinate 'CycleIterator' which
allows to iterate over it in a round-robin fashion, even as its
underlying OrderedMap is mutated.

This will be used, unsurprisingly, in a little TCP load balancer
implementation.

Change-Id: Ic8ab6a7fbcd6e1faf6224fc93bcc997c09364190
Reviewed-on: https://review.monogon.dev/c/monogon/+/1334
Tested-by: Jenkins CI
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>