m/n/core: make startup errors more readable
Previously, fatal startup errors in the cluster manager/enrolment would
get logged as general supervisor runnable errors, which is not very
user friendly.

This reworks error handling in main.go to switch over logging to an
alternative 'fatal' channel as soon as an error is detected, and use it
to clearly log that error.

Fixes issues/77

Change-Id: Id73ee49f6bc5b4306dd67ede3a588f08dd2b33da
Reviewed-on: https://review.monogon.dev/c/monogon/+/1355
Reviewed-by: Leopold Schabel <leo@monogon.tech>
Tested-by: Jenkins CI