m/node/core: run hostsfile from roleserver, provide feedback on cluster directory
Not providing a heartbeat and status until we save a cluster directory
to ESP is a quick and dirty way to make sure we don't mark a node as
HEALTHY until it has performed the bare minimum of setup to be
rebootable.

This is important in our E2E tests to reduce flakiness.

In the future we should have some node status field or general 'sync'
API exposed, but this will do for now.

Change-Id: Ibad9e91f01abeacdfe4400ef7cb36ca17f68ba0a
Reviewed-on: https://review.monogon.dev/c/monogon/+/1498
Tested-by: Jenkins CI
Reviewed-by: Leopold Schabel <leo@monogon.tech>