c/agent: log to extra kernel consoles
Currently the list of consoles to which logtree output is piped is
hardcoded to be tty0 (the first virtual console) as well as ttyS0
(the first platform serial port). On some platforms this is
insufficient, so also log to any kernel consoles defined.

The same problem exists for Metropolis itself, but there we first need a
mechanism to pass a per-machine command line as for safety reasons it's
currently hardcoded in the stub loader.

Change-Id: I918525ed69a87323cbea561f3500b2166c842aae
Reviewed-on: https://review.monogon.dev/c/monogon/+/1511
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>