m/p/supervisor: allow signalling supervised processes
Using Command.Process.Signal to send signals to supervised processes
creates a data race between the Start function (called by Run) which
populates Command.Process and the access to that field.
Introduce a new option for RunCommand which takes a signal channel and
internally solves this race by first calling Start (at which point
Command.Process is populated and will not be mutated) and then spawning
a Goroutine which does the actual sending. This is safe at least on
Linux since Go 1.7 [1].

[1] https://github.com/golang/go/commit/cea29c4a358004d84d8711a07628c2f856b381e8

Change-Id: Ibb76ef8a0434574c4df9ff85545987026518d8c0
Reviewed-on: https://review.monogon.dev/c/monogon/+/1540
Tested-by: Jenkins CI
Reviewed-by: Serge Bazanski <serge@monogon.tech>