c/shepherd/equinix/wrapngo: isolate live API tests
So far all wrapngo tests were fired directly against a live Equinix
endpoint, and the entire test harness was skipped if such an endpoint
was not configured/available.

This consolidates all of the live tests into a single test with
sub-tests, and only that test is skipped if endpoint data is not
provided.

This will in turn allow us to have some unit/regression tests for
wrapngo that do not need to contact the live Equinix API.

Change-Id: I2590a41b935b5d5417520d7510025f6289cece63
Reviewed-on: https://review.monogon.dev/c/monogon/+/1557
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>