m/n/b/mkcpio: work around lib symlink bug
The library we're using to write these, github.com/cavaliergopher/cpio,
does not handle writing symlinks. It does provide a field to put a
symlink target, but it only reads into it and ignores it while writing.
This causes symlinks in our initramfs to be broken. Work around this for
the time being by manually emitting the correct content.

Change-Id: I3e78572d122e802d4a02b48bc87dab5152b2b913
Reviewed-on: https://review.monogon.dev/c/monogon/+/1552
Tested-by: Jenkins CI
Reviewed-by: Tim Windelschmidt <tim@monogon.tech>